# nextjs-build/src/frontend/TestFramework.Client.WebAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TestFramework.WebAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/UITestRuns": {
      "post": {
        "tags": [
          "UITestRuns"
        ],
        "operationId": "PostTestRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUITestRunDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUITestRunDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostUITestRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "UITestRuns"
        ],
        "operationId": "GetUITestRunDto",
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUITestRunDtoPaginationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUITestRunDtoPaginationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUITestRunDtoPaginationResult"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/UITestRuns/{id}/TestCases": {
      "get": {
        "tags": [
          "UITestRuns"
        ],
        "operationId": "GetUITestRunCases",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUITestRunCasesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUITestRunCasesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUITestRunCasesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorCodesResult": {
        "type": "object",
        "properties": {
          "ErrorCodes": {
            "type": "array",
            "items": {
              "enum": [
                "IdLessThen1",
                "UIElements_NameMaxLength",
                "UIElements_NameIsEmpty",
                "UIElements_NameAlreadyUsed",
                "UIElements_FindByIsEmpty",
                "UIElements_FindByMaxLength",
                "UITestCases_NameMaxLength",
                "UITestCases_NameIsEmpty",
                "UITestCases_NameAreadyUsed",
                "UITestCases_StartUrlIsEmpty",
                "UITestCases_UIElementNotFound",
                "UITestCases_UserFilesNotFound",
                "UITestCases_StepsAreNotUnique",
                "UIEvents_NameIsEmtpy",
                "UIEvents_NameMaxLength",
                "UITestRuns_TestCasesNotFound",
                "UITestRuns_TestCasesAreEmpty",
                "UIPages_NameIsEmpty",
                "UIPages_NameMaxLength",
                "UITestCaseStatusInfo_StartIsEmpty",
                "UITestCaseStatusInfo_EndIsEmpty",
                "UITestCaseStatusInfo_EndLessThenStart",
                "HealthChecks_NameIsEmpty",
                "HealthChecks_NameMaxLength",
                "HealthChecks_UriIsEmpty",
                "UserFiles_FileNameIsEmpty",
                "UserFiles_FileSizeIsNegative",
                "UserFiles_FileNameMaxLength",
                "PageIndexLessThan0",
                "PageSizeLessThen1",
                "DeletionFailed"
              ],
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUITestRunCaseDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Start": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Duration": {
            "type": "string",
            "format": "time-span",
            "nullable": true,
            "example": "00:00:00"
          },
          "Browser": {
            "enum": [
              "Chrome",
              "Edge",
              "Firefox"
            ],
            "type": "string"
          },
          "State": {
            "enum": [
              "NotStarted",
              "Started",
              "Failure",
              "Completed"
            ],
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "GetUITestRunCasesDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Start": {
            "type": "string",
            "format": "date-time"
          },
          "State": {
            "enum": [
              "Started",
              "Failure",
              "Completed"
            ],
            "type": "string"
          },
          "TestCases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetUITestRunCaseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUITestRunDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "Start": {
            "type": "string",
            "format": "date-time"
          },
          "CompletedCount": {
            "type": "integer",
            "format": "int32"
          },
          "FailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "StartedCount": {
            "type": "integer",
            "format": "int32"
          },
          "NotStartedCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GetUITestRunDtoPaginationResult": {
        "type": "object",
        "properties": {
          "Count": {
            "type": "integer",
            "format": "int32"
          },
          "Data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetUITestRunDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostResponse": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PostUITestRunDto": {
        "type": "object",
        "properties": {
          "SelectedTestCases": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "required": [
          "TypeDiscriminator"
        ],
        "type": "object",
        "properties": {
          "TypeDiscriminator": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { },
        "discriminator": {
          "propertyName": "TypeDiscriminator"
        }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          }
        ],
        "properties": {
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}